create new file to account_payment_term
