+++ l10n_br_account_product/models/__init__.py
@@ -10,6 +10,7 @@ from . import l10n_br_tax_definition_company_product
 from . import l10n_br_account
 from . import account_product_fiscal_classification
 from . import account_invoice
+from . import account_payment_term
 from . import account
 from . import product
 from . import res_country
